requirements.txt

attrs>=19.3.0
voluptuous>=0.12.0